requirements.txt
Django==1.5.1
bpython==0.12
django-model-utils==1.4.0
logutils==0.3.3
django-braces==1.3.1
django-tastypie==0.11.0
python-mimepars==0.1.4
python-dateutil>=2.2
simplejson>=3.3.3
HTMLParser==0.0.2